telegram_bot/requirements.txt
python-telegram-bot>=20.8,<21.0
python-dotenv>=1.0
peewee>=3.17
psycopg2-binary>=2.9
google-api-python-client
google-auth
google-auth-oauthlib
PySide6